Add `test_helper` module for shared test functions
A number of color space tests reused the same `assert_almost_equal` test
function, which had to be duplicated when tests were moved to separate
sub-modules.  By moving the function to a `test_helper` module, we can
share test functions across sub-modules, and include the helper module
only for test configurations.
